% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/joint_probs.R
\name{get_joint_probs}
\alias{get_joint_probs}
\title{MRmP - Joint Probabilities}
\usage{
get_joint_probs(data)
}
\arguments{
\item{states}{A list comprised of individual dataframes of
state abbreviations using get_joint_margins()}
}
\description{
Creats joint probabilities for a combination of varabiles - derived from their marginal distributions
}
\examples{
x <- get_joint_margins(states = c('DC', 'FL'),
 vars = c('sex', 'age', 'race', 'education'))
get_joint_probs(x)
}

