% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/get_margins.R
\name{get_margins}
\alias{get_margins}
\title{MRmP - State Margins}
\usage{
get_margins(df, states, vars)
}
\arguments{
\item{df}{A data frame with the population margins for each group as column
vector.}

\item{states}{Character vector of state abbreviations. Alternatively or you
can pass it \sQuote{ALL} to get margins for all states in DC.}

\item{vars}{List of variables to get marginal proportions for each state:
sex, age, race, education.}
}
\description{
Gets state level marginal proportions for varyings groups such as age, sex,
education, race
}
\examples{
get_margins(df = data, states = c('DC', 'FL'), vars = c('sex', 'age', 'race', 'education'))
}
\author{
James Wyatt
}

