% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmp.R
\name{mrmp}
\alias{mrmp}
\title{MRmP - Multilevel Regression and Marginal Poststratification}
\usage{
mrmp(survey_data, jointp_list, individualvars, groupingvars, response,
  survey_sample = NULL, weights = NULL)
}
\arguments{
\item{survey_data}{A survey data frame}

\item{jointp_list}{A data frame containg the weighted frequency of each strata}

\item{individualvars}{variables used in mixed effects formula that will have random intercepts}

\item{groupingvars}{variables used in mixed effects formula that serve as state-level grouping vars and do not have random intercepts}

\item{response}{The response/outcome variable on which to regress}

\item{survey_sample}{Total amount to sample from surve_data}

\item{weights}{Weights to be used in the bglmer}
}
\description{
Creates subnational estimates based on the weighted frequence
of each subgroup based on their joint probabilities
}

