% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmp.R
\name{mrmp}
\alias{mrmp}
\title{MRmP - Multilevel Regression and Marginal Poststratification}
\usage{
mrmp(survey_data, jointp_list, mrmp_formula, survey_sample = NULL)
}
\arguments{
\item{survey_data}{A survey data frame}

\item{jointp_list}{A data frame containg the weighted frequency of each strata}

\item{mrmp_formula}{A formulat to fit on survey_data}

\item{survey_sample}{Total amount to sample from surve_data}
}
\description{
Creates subnational estimates based on the weighted frequence
of each subgroup based on their joint probabilities
}

